requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.2.2               # via kombu
argon2-cffi==18.1.0
asn1crypto==0.24.0        # via cryptography
bcrypt==3.1.4
billiard==3.5.0.3         # via celery
boto3==1.7.20
botocore==1.10.20         # via boto3, s3transfer
celery[redis]==4.1.0
cffi==1.11.5              # via argon2-cffi, bcrypt, cryptography
coreapi==2.3.3            # via drf-yasg, openapi-codec
coreschema==0.0.4         # via coreapi, drf-yasg
cryptography==2.2.2       # via pyjwt
customerio==0.4
defusedxml==0.5.0         # via djangorestframework-xml
diff-match-patch==20121119  # via django-import-export
django-ajax-selects==1.7.0
django-braces==1.13.0
django-cors-headers==2.2.0
django-fsm==2.6.0
django-import-export==1.0.0
django-object-actions==1.0.0
django-redis==4.9.0
django-sendgrid-v5==0.6.87
django-simple-history==2.3.0
django-storages==1.6.6
django==2.0.3
djangorestframework-camel-case==0.2.0
djangorestframework-datatables==0.4.0
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.3.0
djangorestframework==3.8.2
docutils==0.14            # via botocore
drf-tracking==1.4.0
drf-yasg==1.8.0
et-xmlfile==1.0.1         # via openpyxl
future==0.16.0            # via drf-yasg
gevent==1.3.1
greenlet==0.4.13          # via gevent
gunicorn==19.7.1
idna==2.6                 # via cryptography
inflection==0.3.1         # via drf-yasg
itypes==1.1.0             # via coreapi
jdcal==1.4                # via openpyxl
jinja2==2.10              # via coreschema
jmespath==0.9.3           # via boto3, botocore
kombu==4.1.0              # via celery
markupsafe==1.0           # via jinja2
mysqlclient==1.3.12
newrelic==2.106.1.88
nexmo==2.0.0
odfpy==1.3.6              # via tablib
openapi-codec==1.3.2      # via drf-yasg
openpyxl==2.5.3           # via tablib
pillow==5.1.0
pycparser==2.18           # via cffi
pycryptodome==3.6.1
pyjwt[crypto]==1.6.1      # via djangorestframework-jwt, nexmo, twilio
pysocks==1.6.8            # via twilio
python-dateutil==2.7.3    # via botocore
python-http-client==3.0.0  # via django-sendgrid-v5, sendgrid
pytz==2018.3              # via celery, django, drf-tracking, twilio
pyyaml==3.12              # via tablib
raven==6.7.0
redis==2.10.6             # via celery, django-redis
requests-toolbelt==0.8.0
requests==2.18.4
ruamel.yaml==0.15.37      # via drf-yasg
s3transfer==0.1.13        # via boto3
sendgrid==5.3.0           # via django-sendgrid-v5
six==1.11.0               # via argon2-cffi, bcrypt, cryptography, drf-yasg, python-dateutil, twilio
tablib==0.12.1            # via django-import-export
twilio==6.12.0
unicodecsv==0.14.1        # via tablib
uritemplate==3.0.0        # via coreapi, drf-yasg
vine==1.1.4               # via amqp
xlrd==1.1.0               # via tablib
xlwt==1.3.0               # via tablib
xmltodict==0.11.0
